+++ lib/Utilities_and_helper_widgets/theme_service.dart
@@ -15,9 +15,10 @@ import 'package:get/get.dart';
 import 'package:get_storage/get_storage.dart';
 
 class Themes {
-  final ThemeData lightTheme;
-  final ThemeData darkTheme;
   Themes({required this.lightTheme, required this.darkTheme});
+
+  final ThemeData darkTheme;
+  final ThemeData lightTheme;
 }
 
 class ThemeService {
@@ -26,11 +27,12 @@ class ThemeService {
 
   ThemeMode get theme => _loadThemeFromBox() ? ThemeMode.dark : ThemeMode.light;
 
-  bool _loadThemeFromBox() => _box.read(_key) ?? false;
-  _saveThemeToBox(bool isDarkMode) => _box.write(_key, isDarkMode);
-
   void switchTheme() {
     Get.changeThemeMode(_loadThemeFromBox() ? ThemeMode.light : ThemeMode.dark);
     _saveThemeToBox(!_loadThemeFromBox());
   }
+
+  bool _loadThemeFromBox() => _box.read(_key) ?? false;
+
+  _saveThemeToBox(bool isDarkMode) => _box.write(_key, isDarkMode);
 }