+++ lib/animations/hero_page_route.dart
@@ -2,6 +2,7 @@ import 'dart:ui';
 
 import 'package:flutter/material.dart';
 
+///USED TO CRAETE A CUSTOM TRANSITON ANIMATION LIKE HERO ANIMATION
 class HeroDialogRoute<T> extends PageRoute<T> {
   /// {@macro hero_dialog_route}
   HeroDialogRoute({
@@ -14,28 +15,16 @@ class HeroDialogRoute<T> extends PageRoute<T> {
   final WidgetBuilder _builder;
 
   @override
-  bool get opaque => false;
+  Color get barrierColor => Colors.black12;
+
+  @override
+  Curve get barrierCurve => Curves.elasticIn;
 
   @override
   bool get barrierDismissible => false;
 
   @override
-  Duration get transitionDuration => const Duration(milliseconds: 500);
-
-  @override
-  bool get maintainState => true;
-
-  @override
-  Color get barrierColor => Colors.black12;
-  @override
-  Curve get barrierCurve => Curves.elasticIn;
-  @override
-  ImageFilter get filter => ImageFilter.blur(sigmaX: 20, sigmaY: 20);
-  @override
-  Widget buildTransitions(BuildContext context, Animation<double> animation,
-      Animation<double> secondaryAnimation, Widget child) {
-    return child;
-  }
+  String get barrierLabel => 'Popup dialog open';
 
   @override
   Widget buildPage(BuildContext context, Animation<double> animation,
@@ -44,5 +33,20 @@ class HeroDialogRoute<T> extends PageRoute<T> {
   }
 
   @override
-  String get barrierLabel => 'Popup dialog open';
+  Widget buildTransitions(BuildContext context, Animation<double> animation,
+      Animation<double> secondaryAnimation, Widget child) {
+    return child;
+  }
+
+  @override
+  ImageFilter get filter => ImageFilter.blur(sigmaX: 20, sigmaY: 20);
+
+  @override
+  bool get maintainState => true;
+
+  @override
+  bool get opaque => false;
+
+  @override
+  Duration get transitionDuration => const Duration(milliseconds: 500);
 }