+++ lib/widgets/text.dart
@@ -1,29 +1,31 @@
-import 'dart:ui';
 import 'package:flutter/material.dart';
 
+///THIS WIDGET IS USED TO CREATE A SIMPLE TEXT WIDGET
 class Txt extends StatelessWidget {
-  final String text;
-  final Color? color;
-  final FontWeight weight;
-  final double size;
-  final TextAlign textAlign;
   const Txt(this.text,
-      {this.color ,
+      {Key? key,
+      this.color,
       this.weight = FontWeight.normal,
       this.size = 16,
-      this.textAlign = TextAlign.center});
+      this.textAlign = TextAlign.center,
+      this.shadows})
+      : super(key: key);
+
+  final Color? color;
+  final List<Shadow>? shadows;
+  final double size;
+  final String text;
+  final TextAlign textAlign;
+  final FontWeight weight;
+
   @override
   Widget build(BuildContext context) {
     return Text(
       text,
-      overflow: TextOverflow.visible,
-      softWrap: true,
+      overflow: TextOverflow.ellipsis,
       textAlign: textAlign,
       style: TextStyle(
-        color: color,
-        fontWeight: weight,
-        fontSize: size,
-      ),
+          color: color, fontWeight: weight, fontSize: size, shadows: shadows),
     );
   }
 }